Updated exception handling for all stol calls

+++ include/process.h
@@ -18,14 +18,14 @@ class Process {
   std::string Command();                   // TODO: See src/process.cpp
   float CpuUtilization();                  // TODO: See src/process.cpp
   std::string Ram();                       // TODO: See src/process.cpp
-  long int UpTime();                       // TODO: See src/process.cpp
+  long long UpTime();                       // TODO: See src/process.cpp
   bool operator<(Process const& a) const;  // TODO: See src/process.cpp
 
   // TODO: Declare any necessary private members
  private:
   int pid_{-1};
   std::string user_, command_;
-  long int uptime_{0};
+  long long uptime_{0};
   float cpuutilization_{0.0};
   ProcMemUtilization* procMemUtilization;
   void CalculateCPUUtilization();
